Renamed Ace to Proof.
Going to see if I can grow accustomed to this new name. I like the name Ace so
much, but there are already two very popular projects in the Node.js realm that
are named Ace. I'm concerned. Given that I've already had someone open a ticket
on Strata, because someone else has project named Strata, I can imagine that
there will be two tickets opened on Ace.

Yet, Ace is perfect for a test framework. If Ace your tests, that makes you an
Ace. It has a Unicode character that I can display in the test output in Travis
CI. It was a lot of fun, which means I might come back to it.

Proof is a good enough name. It doesn't have the right spirit, though. It gets
back into that sort of quality control language. Not as much fun. Has that dour
burden of testing about it.

The name Ace is fun. If I find I'm having less fun with proof, I'll switch the
name back to Ace.

The real drawback is its place in the NPM namespace. It comes up in the search
with the ace framework. That framework is authored by a fellow from Twitter, so
adoption seems to be happening rather quickly. When last I checked, this was
some silly little three line project, with a lot of dependencies. It was a
project to create a file server. It is probably the case that that project has
been replaced by this project, so this person has been working for this name.

I'm not eager to make too much of a test framework. I don't want to have long
discussions about specs and mocking and all those many things for which I do not
care. I don't want too much adoption, then, unless people are content to test
the way I test, with little test programs, until I as the author, have some
confidence in the code.

Thus, proof can be tucked off to the side. With the name ace, I was excited
about this project. With the name ace and naming conflicts, I am not excited
about this project. With the name proof, I am not excited about this project.

I don't much like the idea of being excited about a test framework. I'd be
disappointed if this became something that required a lot of my time, a lot of
my attention, a lot of my care. I'd rather learn more about algorithms and user
experience, than focus on the process of software.

Also, rewrote the README.md, culling all the hand-wringing, and moving all of
the homeless sentiments to the bottom of the page for yet another future purge.