Created `@cleanup`.
The `@cleanup` method accepts a cleanup function that is run immediately, then
run once again at the program exit. It is run immediately to cleanup a failed
previous run. Cleanup functions are meant to be registered early, in order to
perform this defensive cleanup when the test program is in its most stable
state.

The `@cleanup` method is available to both tests and harnesses. It is probably
best to use `@cleanup` in harnesses. It makes tests simpler, sloppier.

For example, each test in a suite could use a temporary directory that the
harness will create for the test, and delete in a cleanup function. When writing
the tests for the suite,  you could put whatever they want in the temporary
directory and know that it will disappear when the test is done.

Updated the README.md Housekeeping section to describe the `@cleanup` function,
and show examples of its usage. This is conceptually simpler than the `$setup`
and `$cleanup` functions that a harness would pass in though a context.

Closes #1.