Synchronous harness definition functions.
This fell out of converting the tests in Stencil from Stencil's current test
framework to Proof. I had created a wrapper around a `deepEqual` assertion,
called `scanEqual`, that would compare some Stencil code against an array of
tokens, after the code had been run through the scanner. It made my tests a
little less verbose to wrap the scanner call in a helper function along with the
assertion call.

To define this function in a Proof harness, I'd need to have access to the
special `deepEqual` in the Proof `Test` class, so that it would emit the
correctly formatted output to standard out. The update that implemented
`@cleanup` did so by calling the harness definition function with the soon to be
run `Test` object instance bound to `this`. The same `@cleanup` function
implementation could then be used in both harnesses and tests.

Now the `@deepEqual` method with its correctly formatted output is similarly
available to the harness definition function. To create my `scanEqual` method I
simply needed to convert my harness from one that defines the harness context
with a context `Object`, to one that defines the harness context with a
callback function.

However, it turns out the harness generation bits of the library were not
expecting synchronous callback functions. The callback was given a `function
(error, result) {}` style callback that it did not itself call.

This commit allows synchronous callback functions for harness definition. It
also starts an `API.md`, where I made a note of this detail, since it is going
to be one detail to many for the `README.md`. Eventually, `API.md` will contain
dry, function by function reference for Proof. For now, I'm dropping notes in it
on implementation details that do not fit the narrative style of `README.md`.

This commit also makes me wonder if context definition using a context object is
necessary. It would be less code, and more concise documentation to have only
the definition function.

The definition function could even provide a context object to populate, making
it much like the `module` object in the `require` module definition framework,
and saving our friends a line of code to create the context object.

It is pretty easy to do `{ one, two, other }` as a return value in CoffeeScript,
however. I'll have to write some more harness definitions functions to decide.