Default harness name, other naming conventions.
Changed the default harness name from `harness` to `proof`. This allows us to
continue to say `require("./proof")` when using harnesses, which is very clever.

Removed the need for a shebang line at the start of the harness. This was used
to suggest a shebang line for new tests when the test generator was run. Instead
we check the file extension of the harness. If it is a Streamline.js file
extension, or if `_` is given as a parameter to the test generator, then
`_coffee` is used as the executable in the shebang line. Otherwise, `coffee` is
used.

Updated `README.md` to remove shebang line requirement.

Renamed `require("./harness")` to `require("./proof")` in all code examples.
Also, fixed a missing shebang in one of the code examples.

In the course of this, I'd forgotten that this is for CoffeeScript. I'd
forgotten that the harness I'm generating is a CoffeeScript file. I'd have to
make a decision to generate JavaScript and Streamlined JavaScript that I've not
yet made. I do look for a possible `.js` or `._js` file extension on harnesses,
but I do not offer the option to generate a JavaScript test program.

Closes #2.