Create Streamline.js aware `@throws`.
Implemented a `throws` assertion. Unlike the rest of the assertions which are
inherited from the `assert` module of Node.js, this assertion is specialized for
Proof and Streamline.js.

The `throws` assertion is Streamline.js aware. It can be called asynchronously.

The `throws` returns the caught exception, so that the test can perform
additional assertions on the exception properties.

Unlike the Node.js `throws`, our `throws` does not accept a validation function.
Additional validation is performed on the returned exception. The CoffeeScript
existence operator makes subsequent assertions easy to express, accounting for
an expected exception that is not thrown.

Added an overview of assertions to the `README.md`.

Fixed some tests that referenced the outgoing Streamline.js executables.

Closes #7.