Sharper summary line.
Created a new summary line for the progress formatter. Instead of having a line
of underscores to create visual cue for the summary, the summary is instead
right justified, with no divider.

Also, tidied up the time formatting. If a test runs longer than ten seconds, the
progress formatter will display the double digit time in seconds.