Fixed PROOF_NO_CLEANUP environment variable.
`PROOF_NO_CLEANUP` wasn't working.  The Proof library will skip cleanup
functions if the environment variable `PROOF_NO_CLEANUP` is set, regardless of
the value set. I'd not tested the implementation. It skipped cleanup, but in
doing so, it also skipped the test's call to `process.exit`. Now setting
`PROOF_NO_CLEANUP` will cause the tidy logic to loop through an empty list of
cleanup functions, which causes it to exit immediately.

Closes #12.