Failed assertions no longer repeat message.
Failed assertions no longer repeat the assertion message as a comment. In my
Node.js `assert` module function wrappers, I was printing the assertion message
passed into the assertion function as the assertion message. To be thorough, I
printed the assertion message of the caught `AssertionError` as a comment.
However, the `AssertionError` message is, of course, the same as the message
passed into the assertion, the assertion message. It was redundant. It is no
more.

Closes #11.