Remove extra call to `npm install`.
I believe the first call to `npm install` is unnecessary. The second call does
it all. We could have two calls, one that reads the `pacakge.json`, the second
that adds `streamline`, but that would be to chatty.

Note that `streamline` is not a build dependency. It is only run if a test
executes one of the `streamline` executables.