Abend when test is repeated for a test run.
Throwing an exception when a user specifies a test program more then once for a
test run. This would throw off the timing and reporting, because we group that
by the name of the test. Each test should be a simple program that tests one
thing, not a complicated program with command line switches or some other way of
changing the programs behavior between invocations.

If you do have repetitious tests, instead of creating a single program with
different modes of operation, create tiny programs that require a common module
and fire off a function with some parameters.

With this in place, there's no need to include the process ID in the raw output,
so this closes #4 as invalid.