Created `errors` to display error output.
Problem with errors is that output can be interleaved, so we need to gather up
the lines of output after a failed assertion, or else the output of other
assertions get interleaved.

The first formatting style that comes to mind would be one that grouped all the
failed assertions under their failed test, but that means waiting for a full
test to load. There are test with a great many failures, one of the automated
tests, like the one in Timezone that tests every clock transition in the world
since the dawn of standardized time. We might run out of memory if a test of
that nature is really broken and really chatty about it.

What we're going to do for a stab at this problem is create a queue, as we do
with progress, and one go at a time. Chances are the queue will be empty. If
there is one long running test interleaved with a quick test, then the quick
test will be done quickly, and the long running test can take over.

If two long running test are interleaved, then we might want to view the tests
one at a time by piping the test through grep to limit the output inspected, or
else piping it through sort so that the rest results are not interleaved.

Related to #5. I'm not sure if I want the progress runner to display errors,
actually. Progress is one thing, errors are another.

The progress runner probably does have to display errors  I'll weave the
`errors` output into it once I'm happy with how it looks.