Set width and width of timings.
Added a `--digits` parameter to `proof progress` and to the default invocation
of `proof`. It sets the number of digits of the time display. This excludes the
decimal point, which is not a digit. The display always has a millisecond
readout. `4` is both the default and the minimum.

Added a `--width` parameter to set the width of the display in characters. Fixed
the `fill` function to return the correct number of fill characters, it was
returning two less than requested.

Closes #16.
Closes #19.