Failure on error exit, using expected tests.
Fixed a bug where tests were being marked as passing because the
condition for passed tests was whether passed matched actual, which
would cause a test to pass even if not all of the expected tests where
invoked.

Now reporting as a failure test programs that exit with a non-zero exit
code.  This covers the case where tests do not even start because the
test will not compile. In this case the expected and passed assertions
are displayed as zeros, and the test is displayed as a failure. It ought
to be pretty obvious that the test wasn't able to do anything.

The failure of a test program due to a non-zero exit code propagates to
the test run summary.

Closes #6.
Closes #8.