Code tidy.
Felt that I was using the word `program` too much, especially when it
was used to describe an event emitted from a program. Thus, I've
replaced the variable name `program` the the name `event` in a lot of
places.

I was using the fantastically ambiguous variable name `record` to
describe events, so that has also changed to `event`.

Renamed the function `styling` to `bar`.