Wrapping assertions in `proof errors` with context.
Wrapping the errors in `proof errors` with context, showing the standard
error and standard output from the last successful test of the plan, to
the next successful test or the exit. This gives more context for
reading debugging messages. The failed assertions still stand out with
the read check mark.

Removed a debugging message that showed when displaying failed tests.

Closes #24.