Convert Cakefile to Makefile.
Gone is the `Cakefile`. Gone is Twinkie. Now building CoffeeScript using
GNU `make`.

If `inotify-tools` are installed, the user can invoke `make watch` to
obtain incremental builds.

Outputs are build using a `make` rules, with a recipe to create a
JavaScript target from a CoffeeScript source, instead of using
CoffeeScript in a recursive mode. This way, `make` will rebuild only the
files that have become out of date according to the `Makefile`.