Implement pipeline scaffolding.
Proof was initially implemented to have `git` argument convention, where
`proof` is the primary executable and the first argument is a command to
`proof`. The ideal was to break up the functions of the `proof`
executable into smaller, UNIX-way programs that would generate and
consume streams of text. The programs would not be individual
executables, but commands implemented by a single `proof` executable.

Within the `proof` executable, I'd implemented a default pipeline as the
default operation when `proof` is invoked without a command. It would
pipe the output of `proof run` to `proof progress`. This pipeline could
also be invoked using the `piped` command.

With this commit, rather than have a single pipeline built into `proof`,
the `proof` executable will execute arbitrary extension programs. When
`proof` is invoked with a command it does not recognize it will search
the search `PATH` for an executable. Given the command `proof example`
Proof will look for an executable with the name `proof-example`.

An extension program can be written in any language, but I'd recommend
`bash`, because its a shell, so its easy to write pipelines.

The default pipeline is now implemented as an extension program named
`proof-default`. It is called if no command is specified, or else if
`proof default` is invoked.

Additionally, I've provided an executable `proof-getopts`, written in
bash, that implements GNU `getopts`, so that you can implement long
options on any target operating system. This is based on the project
[getopt](https://github.com/bigeasy/getopt/blob/master/getopt). That
project as been added as a submodule. The `Makefile` will make an
executable of it and copy it into the `bin` directory.