Extracting context, Step like asynchronous flow.
A test maintains a context. This is the context that is initialized
through the use of a harness. This has been passed in as the first
argument to the callback function. With CoffeeScript, we were able to
use destructuring assignment to get local variables out of the context
object. For JavaScript, we destructure assignment based on the function
parameter names.

We take advantage of the source code representation of the callback
function returned by `Function.toString` to extract and parse the
parameters. Based on the parameter names, we create an argument array
for the callback using the parameter names as keys into the context
object.

There are two special names.

The parameter names `_` and `callback` will receive a callback function
in case the test function wants to handle asynchronous callbacks itself,
without the use of the control-flow built into the Proof library.

A new method has been added to the test object, `store`. This will
generate a callback that will capture the output of a Node.js callback
with the canonical `(error, result)` signature. If there is an error,
the test will bail out.

Proof now provides asynchronous flow control without Streamline.js. The
flow control is tailored for Proof tests, which are brittle and short
lived. No elaborate efforts are made for error recovery in Proof tests,
so the only action taken on error is bail out. The one context would be
crowded after a while in a long running program, but it shouldn't be a
problem in short, specific Proof programs.

The indentation ought to be minimal. The flow control and harness setup
ought to make for short, succinct test programs.

Streamline.js and CoffeeScript are still supported. The application
developer must simply install Streamline.js or CoffeeScript and use
their executables for the shebang line. Streamline.js support is reduced
to a simple branch that detects `_` in the function signature.

I'm eager to see what problems arise from parsing the function
signatures. I'll have to walk this back if there are problems I can't
imagine. If I do have to walk it back, it would be simply to expose the
`context` object as a property of `Test`. That is actually the case now.
`context` is a property of `Test`, but I like the readability of the
parameter extraction.

Closes #39.
Closes #14.