Using JavaScript flows everywhere.
Created the function `cleanup_` to support Streamline.js. The function
`cleanup` will accept a series of functions to run with the asynchronous
flow control.

The `Test` object now runs everything using the `_invoke` queue of
functions. The `Test` object maintains two queues for the `_invoke`
function to work through. The main queue and the cleanup queue. Cleanup
functions are prepended to the main queue so that they are run
immediately after the flow function that added them terminates. They are
also added to the cleanup queue so that they are also run at exit.

A harness is now a function that defines an flow and the flow functions
are simply prepended to the main flow.

Closes #44.
Closes #43.
Closes #41.
Closes #40.