Named cleanup.
Now when a step function is named `"cleanup"` it is pushed onto a second
series of cleanup functions that is run at exit. It is also run
immediately in the order in which it is encountered in the series of
steps, so that cleanup runs when it is encountered, then re-run at
program exit.

Removed the `cleanup` member function from `Test`. The `cleanup_`
function specific to Streamlines.js is no longer necessary.
Streamline.js can just as easily use a named cleanup function.

Closes #48.