Fix named callbacks.
Named callbacks failed due to typo. Added a test for named callbacks.

Also, ensuring that the step function completes before calling the next
step function. Previously, if a callback was created using the
`callback` function and invoked in the same tick, the next step function
would begin running immediately, before the previous one had returned.
This is obviously a problem, especially if the previous step problem
wants to create more callbacks.

Removed Streamline.js as a development dependency. It is a development
dependency. It needs to be present in order to test streamlined tests.

Reintroduced behavior where when an unnamed callback is called with an
object, then the object properties are added to the context.

Closes #57.
Closes #50.