Wait on pending output before exit.
A program can exit before it's `stdout` and `stderr` have drained. When
this is the case, the `proof run` process would emit an exit line, then
emit test or output lines after the exit. The parser will then raise
exceptions.

As of this commit, if `stdout` or `stderr` are still readable, we wait
for them to drain before emitting the exit line.

Closes #72.