Planless tests, monochrome progress.
Now detecting of actual tests exceed expected tests and reporting a test
failure in `proof progress`. Have yet to implement in `proof errors`. #78
Didn't create a test case for either of these cases. #79

Now able to disable colorization of `proof progress` and `proof errors`
using the `-M` or `--monochrome` switch.

Both `proof errors` and `proof progress` will detect if a plan has
exited early as in a compilation error, producing no plan, exited
cleanly producing no plan, or has begun emitting tests without giving a
plan. If a plan is received after a test, it is not interpreted as a
plan, but instead it is interpreted as standard output.

Closes #77. Report failure if actual tests exceed expected tests.
Closes #76. Disable colorization.
Closes #75. Detect tests without plans.
Closes #22. Display compilation error output.