Variable argument say, die alias for bailout.
Both `bailout` and `say` accept variable arguments.

`say` accepts variable arguments just like `console.log` and prints it
out escaped by the commenting character `#`.

`bailout` accepts a message as its first argument, followed by a one or
more objects to inspect using `say`.. If the message is a one liner and
not terribly long, that is the bailout message. Otherwise there is no
bailout message and the string is added to the inspection.

Alternatively, `bailout` will accept a single exception. It will use the
exception message as the bailout message if it is a one liner and not
terribly long. It will print the stack trace using `say`.

Created the `die` alias for `bailout`.