Convert to closure style.
Converted to a closure style, removing the `Test` object. Began to
implement Cadence. Conitnuing to support Streamline.js and CoffeeScript,
but less and less as out dependence on function names is going to make
it more difficult to use CoffeeScript.

Going forward, the series of named or anonymous function that describe
an asynchronous series will be called a cadence in commit messages. The
functions themselves are called cadence functions.

Closes #91.

Now when multiple new cadences are created within a cadence function,
the cadences are run in order following the cadence function.

Closes #95.

The `fail` method will print a `not ok`.

Closes #94.

`UNTIDY` is working again.

Closes #92.

Removed the Streamline.js specific `throws` and `cleanup_` functions.

Closes #93.

Removed the implicit `context` argument that was used with structured
assignment in CoffeeScript tests. All CoffeeScript tests in all other
projects now use the named parameters.

Closes #80.