Build on Windows.
Use the globbing functions in arguable to convert bash wildcard patterns
that Windows cannot understand into lists of matched files.

Created a Windows executable `bin/proof.cmd` to use during testing.

Update close for both bailout and normal exit. Need to ensure that
`stdout` and `stderr` drain before exiting on Windows.

I've got a problem where if I throw the Bailout exception, it can be
caught by a cadence. I'm going to add a long jump capacity to cadence to
abnormally exit from cadence. Maybe the `Bailout` exception is really an
exception thrown from cadence that takes you all the way out of the
cadence function. Maybe it's just `new cadence()`.

Converted `proof-default` from Bash to Node.js. Still like the idea of
helper programs to extend Proof, but Node.js is not much more difficult
than using Bash.

Where you've been using `tee` in the past, you can instead pipe the
output of a program to two different writable streams, one being the
standard input of a subsequent process, the other being a file stream.
If you want to do something like `sort` or `grep`, you can use
[dominictarr/event-stream](https://github.com/dominictarr/event-stream/).

Remove the `getopt` sub-module. We'll encourage people to write
utilities in Node.js. This obviates the `Makefile` as well.

Starting to work on stopping an petering out, instead of just calling
`process.exit`. This means raising exceptions, closing `stdin`. If I
exit immediately on Windows, data doesn't make it to `stdout` or
`stderr`. Got this updated control flow working in the tests, but it is
not pretty.

Converted all of the Bash tests to Node.js. Many are the same test over
and over again, so we might be able to create a function that runs a
test based on some input and expected output file names.

Closes #101.
Closes #99.
Closes #98.
Closes #47.