Add file type suffix to tests.
Linguist assumes that files ending in `*.t` are Perl tests. Linguist is
the program used by GitHub to determine a source file's language, so my
projects on GitHub are listed as part Perl, or primarily Perl if I've
written more tests than library source.

The `*.t` convention was borrowed from Perl, so the confusion comes as
no surprise. The new Proof convention is to use `*.t.ext` where `ext` is
the suffix ordinarily associated with source files of the language used
to author the test.