Progress exits zero for no input.
Progress would exit with a non-zero exit code when no input was
received. The runner emits no output if there is no input. This update
will cause the progress runner to emit no output and exit with a zero
exit code when no input is ever received on `stdin`.

An alternative behavior might be to ensure that the runner always emits
a plan and eof even if no tests are run, so that the progress runner can
be made to be brittle. However, at this point, the special case of no
tests might be better expressed with no output, than with a chatty zero
summary of no tests.

Closes #114.