Report missing tests from errors/progress runners.
Reporting when tests are either missing from the plan, or when there are
more tests than expected, when running the errors formatter or progress
formatter.

The NPM markdown formatter is formatting an extraneous 5 as a heading,
so I've allowed for one line in the `README.md` to be too long.

Closes #121.