`close` instead of `exit`, drop Node.js 0.6.
The executable tests were failing because they were waiting on an `exit`
from `spawn` instead of a `close` event which waits for all of the
output streams to drain. This is new Node.js 0.8 behavior. In Node.js
`exit` would wait for the output streams to drain.

No longer supporting Node.js 0.6 and no longer building using Node.js
0.6 at Travis CI.

Closes #123.
Closes #124.