Set `TRAVIS` variable inside progress test.
We used to pass environment variables into Argauble so that Arguable
would mock them but we don't do that any longer. I suppose we could
still do that as an option, but we didn't here. Thus we need to set the
`TRAVIS` environment variable to `false` for our `progress` test which
will not print carriage returns on Travis CI. We have recorded our
carriage returns in our test fixture output.