Propagate `env`.
Return to propgating the environment variable through Arguable as an
option. This was the intention of options, to let you mock what you need
to mock however you need to mock it. About to see if this too fixes
Travis CI. The previous explicit set of `process.env['TRAVIS']` worked
fine, but this is somehow better.