Create a minimal program runner.
Migrate to new invocation with a minimal program runner, runs the tests.
Recall that I enjoy how Proof will hide all my debugging spew when it
runs its test running, that I'll only see it when there is an exception
and I run an error report, or when I run the test directly.

Considering writing to standard error, which would leave standard out
availble for piping.

Closes #386.