Added ASCII converstion to get char widths.
This is turning out to be somewhat useful after all. It is not about
ASCII escape codes so much as it is about just converting the few
characters we use in our runners. Not sure what I'd do with a project
that made extensive use of ANSI escape codes and UNICODE.

See #387.